refactor: openAPI 사용 로직 리팩토링
#88
